business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetXit.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetXit implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+Z ");
        messageString.append(command.getParameter("XIT"));
        messageString.append("\n");
        return messageString;
    }
}
